Replaced: deleteEntityWithSubscriptions resource has been removed and now this resource is used inside DELETE on app/api/entities/router.ts when deleteSubscriptions is true; and format the code
